requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
slowapi>=0.1.9
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiohttp>=3.11.0
psycopg2-binary>=2.9.0
asyncpg>=0.30.0
